test: UserService Test Code 추가
- User idx int -> long 수정
- DTO @AllArgsConstructor 추가
- null exception handling 추가

+++ src/main/java/neltia/bloguide/api/controller/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
     }
 
     @GetMapping("{idx}")
-    public ResponseEntity<?> getUserDetail(@PathVariable("idx") int idx) {
+    public ResponseEntity<?> getUserDetail(@PathVariable("idx") Long idx) {
         ResponseResult result;
         result = userService.getUserDetail(idx);
         return ResponseEntity.ok().body(result.getResponseResult());
