+++ src/main/java/neltia/bloguide/api/repository/UserRepository.java
@@ -4,7 +4,7 @@ import neltia.bloguide.api.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    User findUserByIdx(int idx);
+    User findUserByIdx(Long idx);
 
     User findUserByUserId(String userId);
 }
